Adjust tiddler_written handling for a more flexible paradigm.
The base class tiddler_written now loops across functions registered
in tiddlyweb.stores.TIDDLER_WRITTEN_HANDLERS (a list) and calls them,
passing the current store object and the tiddler object that was
written (put or delete).

I've tested this with cachinghoster and tiddlywebplugins.whoosher
running together and it works much more properly than what was in
place before. It does require changes in plugin code that wants to
exercise tiddler_written but there's not a ton of that so far, so
is okay.