Introduce the concept of the Tiddlers class.
A Tiddlers is a collection of tiddlers which on the fly
calculates its own last-modified time and maintains a hash
which can be used as an ETag.

This helps to remove redudant traversals of tiddler collections,
in bags, as we move from reading them from the store to writing
them out in a request.

This is just an initial step in several related changes.