More pylint related changes.
The key one here renamed determine_tiddler_bag_from_recipe to
determine_bag_from_recipe. I will do an audit of plugins and what
not to see how much of a problem this will be. Likely it is not
highly used as it is already rather confusing, whatever its name.