Reorder when a the bags in a recipe are checked for 'read'.
In the previous code we would load up the tiddlers and then
from that batch of tiddlers, check to see if we were able to
read them all. This meant we could load up a few thousand
tiddlers and then say, "oh sorry you can't have them" to the
caller. Better to say "you can't have them" before loading
them.

The previous style is an artifact of a different model of
how bag and recipe constraints were going to interact, one
that was deemed too complicated.