Revamp serializer handling on a tiddler PUT.
This is done to allow individual instances to decide what
serializations for which they will accept a Tiddler PUT.
In 1.0.x PUTs are limited to text/plain and application/json
instead of letting the serialization system raise a
NoSerializationError. Everything else is treated as binary
tiddler.

Now, in 1.1.x, if a serialization exists for the provided
content-type and that serialization has an as_tiddler method,
then the content is passed to the serialization and a "normal"
tiddler is filled in. Otherwise it is a binary tiddler.

There's a caveat and a potential bug here: content is decoded
from UTF-8, which might not be correct. We'll cross that bridge
when we get there.