prevented swallowing of ImportError in tiddlywebconfig.py
This should only be relevant for developers, since a deployed TiddlyWeb
instance should never do an import in the config file. However, on
principle, swallowing exceptions is bad, mmkay...

Unfortunately, there seems to be no reasonable way to write an automated
test for this (even sys.path mangling is not sufficient, as tiddlyweb*
has already been imported).

Closes #6