First stab at generating HTTP tests directly from the urls.map
This is done by processing the data structure created by selector
to build up a list of dictionaries in the same form as used by
test/http_runner.py.

Those tests are then sorted and filled out in rather arbitrary
and currently much too baroque ways to get them to run to satisfactory
completion.

The hope is that some of that baroqueness can be be clarified by
doing more declarative work, probably by making the urls.map file
itself more declarative and skipping the selector processing
entirely.

Some caveats:

* it's JSON everywhere
* the challenger uris are skipped
* we do all the PUTs, then all the GETs, then all the DELETEs
* we go from shortest url to longest in each group, so that
  recipes and bags are in place before tiddlers. We go in the
  opposite direction for DELETEs
* the data involved is dead simple so there's no real testing
  of the serializations or policies or such like

The expectation is that this is a tool to make sure things are
running at their very basic level, nothing more than that.