If a tiddler is text but has no serializaton, when we PUT, decode from UTF-8.
This introduces the utility function pseudo_binary which queries a
content type to see if it text based. The use of this function will
expand throughout the system, but needs to be centralized in one
place. It will probably be joined with a binary check as well, but
the details of that are not yet clear.

This also introduces the notion that textual content _must_ be UTF-8
encoded. This was already true, but was not being handled effectively.