Attempt to address some encoding problems in exception messages.
It appears that between 2.5 and 2.6 there is a difference in
a) how unicode is handled in exception messages b) when __str__
is or is not called.

The diff on this commit changes all errors related to the Store,
such as NoBagError, to be subclasses of StoreError, itself a subclass
of IOError. StoreError overrides __str__ to print messages.

The presenting bug related to this was 2.5 only: When requesting
a utf-8 encoded URI for a recipe, the a UnicodeDecodeError could
happen as a result of the stringification of the exception one
step up the stack. This started happening today on tiddlyspace
quite a bit because some UTF 8 recipes and tiddlers started showing
up.

This took absolutely ages to get so it worked for the presenting
problem. I wouldn't be surprised if the change itself is a problem.