lazystore branch to experiment with using less memory
One of the largest bottlenecks in TiddlyWeb occurs when it
needs to work with a large bag containing lots of tiddlers.
Under those conditions the system can slow to a crawl, in large
part because too much memory is used.

This branch work around this, for tiddlers, by only loading a
tiddler from the store in the local scope in which it is needed,
and only having one tiddler loaded at a time.

This _can_ mean more requests to the store, but this has proven
to be fast and scalable, whereas loading a lot of tiddlers in one
request has not (because of memory consumption and resource
contention).

So:

When a filter is run, it is provided with unloaded tiddler
objects. It is the job of the filter to load the tiddler from
the store.

When a serialization uses a collection of tiddlers, it is the
job of the Collection (a Tiddlers object) to reify the tiddler
from the store as it is yielded from the Collection.

The control module is losing its centrality in this model, and
needs more extensive testing before it can be guaranteed correct.
The current tests are insufficient as they try to work without
the store and thus fail oddly. Tests which need work are marked
with an 'x' prefix.

Committing to do some compare and contrast.