slip over store errors when traversing Tiddler collections
This is necessary to avoid a race condition as described in
http://github.com/tiddlyweb/tiddlyweb/issues/#issue/30

It's not clear if this is OK or not, so I think we should
let it stew for a couple of days to get some review.

The concern is that there might be uses of Tiddler collections
who expect them to return the same set of tiddlers that were
put in.

It should be noted, though, that this special behavior only shows up
when the collection has been created with a store (which is only
sometimes).