Add a title attribute to Collections.
This begins the process of making is possible to state a title
for a serialization of tiddlers (see html.py for example). The
motivation for that is better titles on atom feeds (the work
to update the atom serialization to use this functionality is
not yet done).

To add an arbitrary title, put a "title=some title you like" query
string, on a URI that returns a collection of tiddlers.

These changes also fix bugs:

* with the title of search results
* with the title of revisions results, when a filter is used
* with the presentation of the representation links at the top
  of a list of tiddlers (this list of tiddlers as...)

And it removes a lot of processing within the tight loop in the
HTML serialization's list_tiddlers that should speed things up a
bit.