Experiment with a fix or tiddler collection etags.
Without this change, the modified time for a tiddler collection is
always now(), which means that etag handling fails to engage,
which is rather unfortunate in some cases.

However, the flip side is that this requires all tiddlers being
added to the collection to have been or be read from the store.