fix #19
When a tiddler is created with its name as just ' ' (one single space)
it was unclickable in the HTML serialization because the tiddler.title
becomes plain whitespace, which is ignored in the html.

This change makes all spaces in tiddler.title into an en space: &ensp;

This makes the whitespace visible while allowing linebreaks if the
title needs to flow. The ticket suggested using nbsp, but that would
require checking the tiddler title for a special case. This solution
is simpler: no conditional required.