Ensure that a password which exists but is the empty string cannot auth.
I could have sworn this is a regression but I can find no evidence of that.
I guess it has been wrong for some time.

People who are using the user database to store roles for external
users, such as openids, should upgrade to include this version, and
recreate the users.