Make the tests work with py.test 2.x
Newer py.test has a collection mechanism which runs more code
than the previous collection mechanism did. This means that when
a global like 'config' is used, it can get messed up such that
it has things in it that aren't deepcopy-able.

This addition removes the selector object from config if it
is in there before attempting a deepcopy.