Enable saving, reading and list tiddlers that start with '.'
In the previous code the tiddlers were being saved, but
list_bag_tiddlers could not see them because _files_in_dir was
ignoring anything that started with '.'.

Now it does not.

To make that work without conflicting with the write lock system,
list tiddlers now must make sure that the files it accepts are only
directories (tiddlers) not files (write locks).

This fix inspired by:

http://groups.google.com/group/tiddlyweb/browse_frm/thread/37ffff8153075b1a#