Add a 'render' option to the JSON serialization.
When retrieving one or a collection of tiddlers, if the query
string includes 'render=1' then a new key called 'render' may
be added to the returned tiddler dictionary. The value is the
text of the tiddler sent through the wikitext rendering system.

The render key shows up on a single tiddler only if the tiddler.type
is none or a member of wikitext.render_map.

For a tiddler collection to get render, fat=1 must also be set.
I considered making render imply fat, but explicit is better than
implicit.