Adjust how store is handled with Tiddlers collections
This is an extension to the previous commit (which included more
than it was supposed to).

The change here and in the previous commit attempts to lessen the
number of times a tiddler is pulled from the store in any given request.
It's possible for a request of a list of tiddlers--with a filter, with
an interesting serialization--for each tiddler to be requested more than
once.

The main logic is in code that calls send_tiddlers: if we know we
are going to have filter processing later, we don't need the current
collection to have a store, otherwise we do, because that's how we
calculate the last modified time on the collection.

These changes introduce the possibility for a serialization, such as
html or text which only lists tiddler titles to not need to load
the tiddler from the store to do its presentation. This is done
by setting tiddlers.store = None at the start of the list_tiddlers.