Implement a first spike at a concept known as a remote bag.
on a branch, needs review, has caveats

This is an exploration for an idea to use with a federated tiddlyweb.

The basic idea is that it is possible to make a recipe which contains
remote uris where a bag would normally be listed. When the recipe is
processed the tiddlers that are at the remote end are processed into
the recipe's tiddlers, filtered as things normally would be.

This implementation has been done to be more generic than that, as
it just makes sense. The result of this:

* You can put _any_ URI in a recipe and that URI will be treated
  as a bag containing one tiddler. The tiddler's type will be set
  by the content-type of the resource at the URI.
* A config setting, remote_handlers, can be extended with regular
  expressions and a two-tuple that point at functions that will
  retrieve and process the content at the URI so it looks like
  a collection of tiddlers or a single tiddlers (think processing
  an rss feed, twitter timeline, etc).

The thing on the remote end is treated as a read only bag.

Concerns and caveats:

* There's no auth, so you can't get private stuff, but one can imagine
  ways of hooking it in (oAuth, cookie proxying, the usual)
* The text store does not really lend itself to this stuff, because
  of the way it stores bag info. The main symptom of this is that using
  a uri with a query string will not work: the query string will be treated
  as a filter.
* Remote bag handling is snuck in at the level just above the store,
  which means that you get a wacky side effect. When the feature is
  in place you can do something like:

  http://0.0.0.0:8080/bags/http:%2F%2Fexample.com%2Fanything/tiddlers

  and get a link to the HTML at http://example.com/anything. If you
  then change anything to foobar, you'll get whatever is at foobar.

On the one hand this stuff is really quite awesome and powerful and
another hand it is way more than enough rope to hang you, your friends,
and total strangers too.

Not quite sure how to manage that.

The code is also a bit of mess. Comments, reviews, etc needed.