Further abstractions of the specialbag concept.
config['special_bag_detectors'] is a list of functions which take
a environment and bag name. That function can return None or a tuple
of two functions, one that returns the tiddlers in a bag, and one that
returns a tiddler from that bag.

This will allow the remotebag stuff to be removed to a plugin, one that
can activate itself simply by adding 'is_remote' to 'special_bag_detectors'.

While it is assumed that most special bags will be 'remote' that may not
always be the case so this allows the hooks.