Make it possible to subclass Recipe, Bag and Tiddler and still have stores and serializers work.
Stores and serializers determine what method they are to use when provided
an object by looking at the __class__ of the object. This prevents subclassing.
By using mro()[-2] we take the class that is at the top of any hierarchy,
getting the same results while still allowing subclasses.

For this to work generically, Recipe is no longer a subclass of list,
which it wasn't really getting much benefit from anyway.

Subclasses are handing for remotebag handling of tiddlers.

Committing for the sake of review.