Add a links attribute to Collections.
this makes it possible for callers which establish Collections to
state the link they want to show up in representations, rather than
having that link be guessed by context.

In addition to think making some things easier in plugins, it is
also explicit instead of implicit, which is a goal.

Also various pep8 fixes.