refactor etag handling for single entities to make them more cacheable
* move etag generation into tiddlyweb.web.util
* remove mime-types in etag generation: we Vary on Accept
  so caches will adapt correctly (by including Accept in cache
  identifiers)
* use the JSON serialization to create the string from which a
  hash is created

This last is the major change. For bags and recipes the hash
already came from a pseudo serialization (python repr strings)
so using JSON allows things to be more correct and have less
duplication. For tiddlers the text attribute is set to null before
the serialization is done and we get benefits:

* server.perms are automatically included in the hash digest,
  meaning we can have different etags for different users without
  needing to refer to the indvidual user. This means that two users
  with the same permissions will get the same etag, reducing the total
  number of etags out there, increasing the chances of cache hits.