fix #43 Add 'uri' as a field to the JSON serialization of a tiddler.
this uses tiddler_url to generate the absolute uri for the
tiddler

this makes life a bit easier for the user-agent when wanting to
play games with the tiddler

This is what might be called a developer optimization. It is not
for the sake of better HTTP. For that we might consider adding a
Content-Location, but that's a different set of things.

Much of the diff here is with updating tests to make sure that
the Serialization has a proper environ with config['server_host']