Implement simple manual caching of single tiddlers.
If a tiddler has the field _cache-max-age its value is used to set
the header:

   Cache-Control: max-age=<the value>

overriding the default of:

   Cache-Control: no-cache
   (which actually means you can use the cache if you revalidate the etag
   against the server)

The idea here is that tiddlers which are used often but changed rarely
should be cacheable by the browser. In the TiddlySpace environment this
can mean things like chrjs and jquery aren't rechecked every single time
a single tiddler page is loaded.

The reasons for doing it per tiddler are because:

* granularity is good
* it's most flexible and powerful
* it maps to the tiddly way of thinking about stuff

Other options were doing it bag based, but bags don't have easily
accessible attributes for doing such controls.

Submitted for review and acceptance.