Establish TypeError as the exception when unable to store.
At various levels of the stack, an incoming representation may
not be able to be stored. This is most common with tiddlers,
in particular tags in JSON. If we assume this badness will cause
a TypeError, we can catch that higher up, and deal with it
appropriately.

For now this change only traps tiddler PUT. Waiting on real
world data to properly correct bags and recipes.