First pass of test driven porting to Python3
Most of this is dealing with three things:

* print -> print()
* except Foo, exc -> except Foo as exc
* various bytes, str, unicode weirdness

The remaining tests that are not passing, there are many, are hung
up on encoding issues in wsgi-intercept, which uses a lot of StringIO
and needs to use BytesIO. I've started a port of that too.

Oh and I ported selector as well, will need to wrap that up somewhere
too.

bwains.