More tests passing, including some web tests.
Most of the changes are actually in the ongoing work on wsgi-intercept
and html5lib. wsgi-intercept is now using BytesIO throughout instead of
StringIO, which meant that some of the socket reading that was being
done in the tiddlyweb code needed to be tweaking back to what is
was _before_ the previous commit.