More test updates.
Now stuck on encoding problems within Etags.

I recall this was a challenge in python 2.7 as well.

This may be an issue in my changes to wsgi-intercept which
may be overquoting, need to look more closely at what is
coming through the raw environ, everywhere in the stack.

At the moment not enough brain power.