Host our own version of 'quote'
The need for this comes from a change in urllib.parse.quote.

In python 2.x there was a module level string of "always_safe"
to which a caller could add strings which should be always
safe when URL quoting. This made it possible to more closely
hew to the RFC. This access is gone in python3 and the similar
data structure is "frozenset" so not to be messed with.

So here we simply call our quote, modify the safe parameter
with our local always safes.