Ensure charset in Content-Type headers.
For textual, but not serialized, tiddlers, the mime-type gets
set to just tiddler.type, but in order for browsers to not
moan the charset needs to be stated explicitly. We dictate
utf-8. There is no other option.

Also updated some related tests to be be less noisy in their
assert statements.