fix #53 non-existent serialization extension no longer returns default
This may not be the best fix, but it has the least impact in terms
of changes (despite the size of the diff).

The conceptual change here is that something which is a "collection"
has a strict name "bags", "recipes", "tiddlers", "revisions" and
thus anything '.*' at the end _must_ be an extension and if that
extension is not a member of `extension_types` we should 415.

To accomplish this, when calling tiddlyweb.web.util:get_serialize_type
we pass `collections=True` to engage different code.

This revealed a bug in extension handling for revisions which was
fixed by adding yet _more_ cyclical complexity to `_determine_tiddler`
a method that's ripe for refactoring (now that we have a few years
of use).

This commit brings about a change in behavior, so will be part of
the 1.3.x release jump. The change is that a "bad" extension is
treated as a bad one, on collections, even in the presence of a
valid accept header. In the past we'd let the accept header take
over. This leads to confusion, as described in #53, thus the change.