Connection and improve handling of _canonical_uri field
If a tiddler has a field named _canonical_uri that field's
value is used when creating tiddler_url. In the past two things
were wrong with this:

* that url was not encoded to utf-8: field values are unicode
* that url was only used when presenting a tiddler collection
  when asking for the tiddler itself directly, you'd get an
  empty tiddler

Both are fixed.

Now, if you ask for a tiddler with a _canonical_uri field
in anything but "Accept: application/json" the response
will 302 to _canonical_uri.

This allows the original desired impact, which is to be
able to have meta tiddlers that point to binary content
elsewhere. See this thread:

  http://groups.google.com/group/tiddlyspace/browse_frm/thread/66792d8989139afd

I guess I just forgot to finish it.