Avoid a rare case when content may be decoded twice
It's possible to create a serializer that has no
'as_tiddler' method. When such a thing happens and
a tiddler is PUT, there are circumstance where the
code would be decoded once to test to see if the
serialization can make a tiddler, and then again
when it is processing a psuedo binary tiddler.

Now, we break the abstraction slightly. This is
a hack just as the previous code was a hack, but
this hack is less costly.