Some cleanups in handler/tiddler
* use a constant where _canonical_uri is used
* don't return revision from _check_validate_tiddler, not needed
  and wrong

The value on tiddler.revision at that point is the value of
what is currently in the store not the current new tiddler,
and when we PUT the tiddler it will be set to whatever the
store says it should be. From the store's perspective the
incoming revision is not important, this is just some tiddler
to be put.