Get rid of HTMLPresenter
Do this by moving the header and foot creation into the HTML
serialization.

In the process claim the HTML is HTML5 and declare it utf8.

This removes a tiny bit of functionality in favor of something
that makes considerably more sense for modern use.

This can't be released until doing a check of plugins that
might use the old functionality. tiddlyhoster is one that definitely
does, there may be others. tiddlyspace imports it, but only
to turn it off.

Anything that uses tiddlywebplugins.utils:entitle will require a look,
can probably just change that method rather than changing the things
that call it.

There will also be documentation that needs to be updated.

Would appreciate a gander from @fnd. As you'll recall this is
""let's do the refactoring, deferring changes in the HTML structure to
a separate step [which might or might not happen]". Of special
interest may be the changes in the root handler.