refactor logging to allow syslogging
This switches the logging infrastructure to be more
like what might be expected in, but does not go as far
as being fully configurable.

There are two overarching changes here:

* If log_syslog is set in config, it is expected to be a
  a syslog facility and will cause logging to be sent to
  that facility (implying configuration in syslog.conf like this:

  local7.*                                     /var/log/local7.log

* There are now two loggers and they have named "tiddlyweb" and
  "tiddlywebplugins". Modules and plugins that wish to log should
  put, somewhere in the module, the following:

    import logging
    LOGGER = logging.getLogger(__name__)

  at which point they will then be able to:

    LOGGER.debug('some debug info')

  and it will go to the logging configured in tiddlywebconfig.py.

  The use of two names enable to the use of __name__ above in a
  hierarchical way: You get to know which module the message comes
  from.

This of course means that all the existing logging in plugins
will need to be updated. However not updating will not cause
catastrophic failure: it will just mean lack of logging.