Simplify the calcuation of the container for a collection of tiddlers
Fix #62

The general idea here is a Tiddlers Collection can now have a
bag or recipe attribute which contains the name or the bag or
recipe that is the container for these tiddlers. They default to
None.

This is _very_ useful information in serializations, making it so
that you don't have to appeal to the contained tiddlers or the
wsgi routing args to figure out what we're. Which saves processing
and guesswork.