A (fully) failing test for 304 handling
This tests those resources which might can return a 304 to confirm
that the response includes all those headers which httpbis says
it must (or should):

   http://tools.ietf.org/html/draft-ietf-httpbis-p4-conditional-22#section-4.1

In our case these headers are listed in RELEVANT_HEADERS in the
test file. We don't include content-location because we aren't
using that at the moment.

This if part of issue #77

This fails at the moment, as expected. Now to make the code
align with the tests.

/cc @pads @fnd