Address a rare 500 in list_bag_tiddlers
If a special bag is accessed directly, but for some reason
it's retriever fails a SpecialBagError is raised and then
reraised as a NoBagError. This NoBagError was not being
caught and would result in a 500, instead of the desired 404.

Now it is becomes a 404.

Also changed some error messages from the specific "remote"
messages to a more general "special".