fix #80 json response should media type should not include charset
One change in config, rest is updating tests to reflect. zomg, dryness.

The one place where there still remains some ickiness is with the
handling of pseudo binaries tiddlyweb/web/handler/tiddler. It
automatically adds charset. As these mostly apply to text or
xml forms and these _do_ accept charset we'll leave as is for now.