fix #81 deny storing a tiddler using reserved fields
If a tiddler has fields with key names that duplicate the
core attributes of the tiddler, reject the tiddler with a
TiddlerFormatError. When this was not done, the tiddler
would be saved to the text store in a way that caused
subsequent retrievals to fail.

Issue #81 describes some other ways that this could have been
done. This is the most pragmatic at this time.