fix #86 by adjusting RequestLogger to RequestStarter
Now in addition to starting a debug log at the beginning
of a request it will also replace the provided PATH_INFO
with one calculated from REQUEST_URI if REQUEST_URI is
present. This leaves us with PATH_INFO in the form we
expect it: How it was sent to the server.

This removes the need for twp.pathinfohack but will not
break things if it is not removed from an instance configuration.

review from @fnd and @pads desired (however this is pretty much
exactly what we agreed upon).