Gain minor efficiencies in tiddler etag generation.
Instead of copying tiddler.text to a temporary variable,
and then making a json version the textless tiddler, and
then copying text back, make a new tiddler borrowing some
of the attributes of the tiddler.

For large tiddlers this is a savings.

It also means that tiddlers that lazy load text (as in
the shunted binary experiments) may be able to work more correctly.

Committing on a branch for review from @FND and @pads.

The reason for having such a complex etag in the first place
rather than just revisions, is because we need^Wwant workable etags
for tiddlers kept in stores that don't use revisions.