Make web_tiddler tests go.
One notable changed here, amongst the usual encoding stuff, is in
the JSON serializer. It makes sure that even when text is '' you'll
still get some output _and_ b64encode won't explode (it requires a
bytes not a str). This is important when you say you want 'fat', you
expect the `text` key to be present even if empty.