Remove a bogus encoding test.
The only reason the provided query string is getting through
is because of wsgi-intercept and the test setup. Under any
normal circumstances it would not. Since the test has to be
ignored under python3 anyway, and works on some python versions
and not others, just killing it. There is potentially a subtle
bug in the SimpleLog middleware, but this test is not really
dealing with that, so is effectively noise.