Never accept incoming modifier on tiddler PUT
In the past, because of TiddlyWiki classic, a decision was made
such that if the current wiki was not using auth, then when
PUTting a tiddler (and the current user was GUEST) the tiddlywiki's
own notion of the user would be set as the modifier.

This turns out to be a significant error, as described in the
comments linked to the issue. This fixes #97.

This leaves some of the tests being a bit redundant (in
test_web_tiddler_revisions), but I felt it better to limit
the area of change here. Certainly in the two tests modified
there are _lots_ of things that could be improved in terms of
interacting with the API. I suspect many of these tests were
_very_ early.