Experiment with also reading query data from multipart
If any of the form items are files, add then to a
tiddlyweb.input_files environ variable so they could be read
later. We do _not_ want to read them automatically.

May kill this enitrely, but trying it because if you want to
csrf while also allowing file uploads over post, this is needed
(or big changes elsewhere).