Add explicit coverage from bag and recipe_etag methods
These utility methods are not used by the core code because
common code was refactored to support both bags and recipes.

However the utility methods have good names, so we want some
tests in there to make sure they are happy.

The tests simply assure that the etags of two identical bags are
the same.