Add 'application/x-tiddler+json' type alias for JSON serialization
This is for #99 but may not yet be a complete fix.

What happens here is that the JSON serializer gets used when
media type is either application/json or application/x-tiddler+json

What's different is that using the latter, on a GET, makes it possible
to retrieve a tiddler with .type as application/json as a full
tiddler structure.

To make that work the login in get_serialize_type was adjusted
so that it (optionally) returns the accept type that matched for
the serialization chosen. Then in the tiddler get handler, we can
compare that accept type against the type of the tiddler. If those
are the same, we send as a binary or pseudo binary. If they don't
we send as a tiddler. The previous code check on the _outgoing_
media type for the chosen serialization.

Note that the outgoing media type remains application/json as this
is what many APIs seem to do. I suspect because it eases content
handling on the client side.

Look at the tests to see what's really going on.

One main remaining issue:

* As a side effect of the configuration change done to make the above
* work, you can now get other JSON reps as JSON by asking for
  x-tiddler+json. This is a bit semantically weird. This suggests,
  perhaps, that the incoming Accept header should be one of:

  * x-tiddlyweb+json
  * vnd.tiddlyweb+json

  or something of that ilk? Thoughts @FND, @Jermolene, @pmario?