Add support for application/x-tiddler-dictionary type
Used by tiddlywiki5 as a fielded textual data representation.

This is based on https://github.com/tiddlyweb/tiddlyweb/pull/102
from @andrey013 but done slightly differently to make room for future
expansion[1] and to have tests.

Note that presumably this could have been done as a tiddlywiki5
related plugin but I chose not to in this case because I want to
reward tiddlywiki5's capability of using an extremely limited
tiddlyweb server and still get stuff done. The only other plugin
required at this point is tiddlywebplugins.status to handle user
info. This is a vast improvement over tiddlywebwiki.

[1] It might make sense in the future for the pseudo binary types to
be configurable from tiddlywebconfig.py, but YAGNI for now.