Update yield generate tests to use pytest_generate_tests
pytest has deprecated yield style generated tests and emits warnings
when they are used. These can be replaced by using the
pytest_generate_tests method which is one of several mystical methods
that are provided in modern pytest. See this thread (it starts before
october):

    https://mail.python.org/pipermail/pytest-dev/2016-October/003905.html

This process does essentially the same thing but it is done in a way
that is more correct with how pytest does test collection and per
test fixture handling.

These changes allowed (or revealed) some cleanups to the imports and
removal of some globals.