Initial work to get tox running
As part of a process to refresh the tiddlyweb tests to use
modern py.test (which no longer support yield tests) start
using tox to run the tests. The provides easier management
of dependencies and multi-version testing. This gets py27,
py37, pypy, and docs working, but not yet pep8.

Three tests which relied on the ordering of dicts have been
removed as they don't really provide any additional coverage
_now_. They were useful during development.

extras_require is used to create dependency management without
requiring requirements.txt and test-requirements.txt (which are
supposed to be used for apps, not libraries).

warnings as errors are turned off because we don't want those
active while we are proving the testing system.